Add picture package to :data2
