+++ app/build.gradle.kts
@@ -126,11 +126,9 @@ dependencies {
     implementation(project(":feature:onecontactwidget"))
     implementation(project(":feature:contactconfig"))
     implementation(project(":feature:settings"))
-    implementation(project(":data:contacts:api"))
     implementation(project(":data:contacts:impl:default"))
     implementation(project(":data:phones:api"))
     implementation(project(":data:phones:impl:device"))
-    implementation(project(":data:pictures:api"))
     implementation(project(":data:pictures:impl:default"))
     implementation(project(":data2"))
     implementation(project(":core:ui"))
@@ -201,7 +199,6 @@ dependencies {
     testImplementation(libs.kotlinx.coroutines.test)
     testImplementation(libs.hilt.android.testing)
     kaptTest(libs.hilt.android.compiler)
-    testImplementation(project(":data:pictures:impl:mock"))
 
     androidTestImplementation(libs.androidx.test.core)
     androidTestImplementation(libs.androidx.test.runner)
