--- app/src/main/kotlin/com/blaxsoftware/directcallwidget/domain/SaveOneContactWidgetUseCase.kt
@@ -1,55 +1,0 @@
-/*
- * Direct Call Widget - The widget that makes contacts accessible
- * Copyright (C) 2025 Fer P. A.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package com.blaxsoftware.directcallwidget.domain
-
-import android.graphics.Bitmap
-import android.net.Uri
-import dev.ferp.dcw.core.di.IoDispatcher
-import dev.ferp.dcw.core.domain.data.PhoneType
-import dev.ferp.dcw.core.domain.data.onecontactwidget.OneContactWidgetRepository
-import dev.ferp.dcw.data.pictures.WidgetPictureRepository
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import javax.inject.Inject
-
-class SaveOneContactWidgetUseCase @Inject constructor(
-    private val widgetRepository: OneContactWidgetRepository,
-    private val pictureRepository: WidgetPictureRepository<Uri, Uri, Bitmap, Int>,
-    @IoDispatcher private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO
-) {
-    suspend operator fun invoke(
-        appWidgetId: Int,
-        displayName: String?,
-        phoneNumber: String,
-        phoneType: Int?,
-        selectedPictureUri: Uri?
-    ) = withContext(ioDispatcher) {
-        val internalPictureUri = selectedPictureUri?.let {
-            pictureRepository.addPicture(it)
-        }?.getOrNull()
-        widgetRepository.addWidget(
-            appWidgetId = appWidgetId,
-            displayName = displayName ?: "",
-            phoneNumber = phoneNumber,
-            phoneType = PhoneType.UNKNOWN,
-            pictureUri = internalPictureUri?.toString() ?: ""
-        )
-    }
-}
