+++ data/pictures/impl/default/build.gradle.kts
@@ -36,7 +36,6 @@ android {
 }
 
 dependencies {
-    implementation(project(":data:pictures:api"))
     implementation(project(":core:di"))
     implementation(project(":core:util"))
     implementation(project(":core:domain"))
