--- data/pictures/impl/default/src/test/java/dev/ferp/dcw/data/pictures/DefaultPictureRepositoryTest.kt
@@ -1,104 +1,0 @@
-package dev.ferp.dcw.data.pictures
-
-import android.content.ContentResolver
-import android.content.Context
-import android.graphics.Bitmap
-import android.graphics.Bitmap.Config
-import android.graphics.BitmapFactory
-import androidx.core.net.toFile
-import androidx.core.net.toUri
-import androidx.test.core.app.ApplicationProvider
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.google.common.truth.Truth.assertThat
-import dev.ferp.dcw.core.util.test.MainDispatcherRule
-import dev.ferp.dcw.data.pictures.rules.ImageRule
-import dev.ferp.dcw.data.pictures.source.disk.PictureLoader
-import io.mockk.coEvery
-import io.mockk.mockk
-import kotlinx.coroutines.test.StandardTestDispatcher
-import kotlinx.coroutines.test.TestDispatcher
-import kotlinx.coroutines.test.runTest
-import org.junit.Before
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-import java.io.File
-
-// TODO Review these tests
-@RunWith(AndroidJUnit4::class)
-class DefaultPictureRepositoryTest {
-
-    private val dispatcher: TestDispatcher = StandardTestDispatcher()
-
-    @get:Rule
-    val mainDispatcherRule = MainDispatcherRule(dispatcher)
-
-    @get:Rule
-    val imageRule = ImageRule()
-
-    private lateinit var context: Context
-    private lateinit var contentResolver: ContentResolver
-    private lateinit var picturesDir: File
-    private lateinit var pictureLoader: PictureLoader<Bitmap>
-    private lateinit var repo: DefaultPictureRepository
-
-    @Before
-    fun setUp() {
-        context = ApplicationProvider.getApplicationContext()
-        contentResolver = context.contentResolver
-        picturesDir = File(context.filesDir, "pics")
-        pictureLoader = mockk()
-        repo = DefaultPictureRepository(
-            contentResolver, picturesDir, pictureLoader, dispatcher
-        )
-    }
-
-    @Test
-    fun `addPicture copies file from uri to pictures dir`() = runTest {
-        // Given
-        val sourceImage = imageRule.createSampleImageFile()
-
-        // When
-        val addPictureResult = repo.addPicture(sourceImage.fileUri.toString())
-
-        // Then
-        assertThat(addPictureResult.isSuccess).isTrue()
-        val resultFile = addPictureResult.getOrNull()?.toUri()?.toFile()
-        assertThat(resultFile?.exists()).isTrue()
-        assertThat(resultFile?.parentFile).isEqualTo(picturesDir)
-        val resultBitmap = BitmapFactory.decodeFile(addPictureResult.getOrNull()?.toUri()?.path)
-        assertThat(resultBitmap.width).isEqualTo(sourceImage.width)
-        assertThat(resultBitmap.height).isEqualTo(sourceImage.height)
-        assertThat(resultBitmap.config).isEqualTo(sourceImage.config)
-    }
-
-    @Test
-    fun `getPicture returns the right picture`() = runTest {
-        // Given
-        val expectedBitmap = Bitmap.createBitmap(100, 100, Config.ARGB_8888)
-        coEvery {
-            pictureLoader.loadPicture(any(), any(), any(), any())
-        } returns expectedBitmap
-
-        // When
-        val getPictureResult = repo.getPicture("fakeUri", 200, 200)
-
-        // Then
-        assertThat(getPictureResult.isSuccess).isTrue()
-        assertThat(getPictureResult.getOrNull()).isEqualTo(expectedBitmap)
-    }
-
-    @Test
-    fun `deletePicture deletes the file from disk`() = runTest {
-        // Given
-        val sourceImage = imageRule.createSampleImageFile()
-        val addPictureResult = repo.addPicture(sourceImage.fileUri.toString())
-        assertThat(addPictureResult.isSuccess).isTrue()
-
-        // When
-        repo.deletePicture(addPictureResult.getOrNull()!!)
-
-        // Then
-        assertThat(addPictureResult.getOrNull()?.toUri()?.toFile()?.exists()).isFalse()
-    }
-}
