--- data/pictures/impl/mock/.gitignore
@@ -1,1 +1,0 @@
-/build
