+++ settings.gradle.kts
@@ -37,13 +37,10 @@ include(":app")
 include(":feature:onecontactwidget")
 include(":feature:contactconfig")
 include(":feature:settings")
-include(":data:contacts:api")
 include(":data:contacts:impl:default")
 include(":data:phones:api")
 include(":data:phones:impl:device")
-include(":data:pictures:api")
 include(":data:pictures:impl:default")
-include(":data:pictures:impl:mock")
 include(":data:onecontactwidget:source:sharedprefs:impl:default")
 include(":data:onecontactwidget:source:sharedprefs:impl:mock")
 include(":data2")
